# bff-api/docs/swagger.yaml
basePath: /
definitions:
  models.Contact:
    properties:
      contactMessageCount:
        type: integer
      identity:
        $ref: '#/definitions/models.Identity'
      totalMessageCount:
        type: integer
      yourMessageCount:
        type: integer
    type: object
  models.ContactTimeline:
    properties:
      identity:
        $ref: '#/definitions/models.Identity'
      timeline:
        items:
          $ref: '#/definitions/models.DayCount'
        type: array
    type: object
  models.DayCount:
    properties:
      count:
        type: integer
      date:
        type: string
    type: object
  models.Group:
    properties:
      archived:
        type: boolean
      creator:
        type: string
      group_member:
        items:
          $ref: '#/definitions/models.GroupMember'
        type: array
      group_name:
        type: string
      group_uid:
        type: string
      id:
        type: string
      message_count:
        type: integer
    type: object
  models.GroupMember:
    properties:
      identity:
        $ref: '#/definitions/models.Identity'
      message_count:
        type: integer
    type: object
  models.GroupTimeline:
    properties:
      group:
        type: string
      identity:
        $ref: '#/definitions/models.Identity'
      timeline:
        items:
          $ref: '#/definitions/models.DayCount'
        type: array
    type: object
  models.Identity:
    properties:
      firstName:
        type: string
      identity:
        type: string
      identityID:
        type: string
      lastName:
        type: string
      nickName:
        type: string
    type: object
  models.WordCount:
    properties:
      count:
        type: integer
      word:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: BFF API
  version: "1.0"
paths:
  /avatar/{id}:
    get:
      description: Gets Avatar for CONTACT, Group or Avatar by ID
      parameters:
      - description: CONTACT, GROUP or AVATAR
        enum:
        - CONTACT
        - GROUP
        - AVATAR
        in: query
        name: type
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: Image
          schema:
            type: file
  /contacts/:
    get:
      consumes:
      - application/json
      description: Fetches all contacts along with their message counts from the database
      produces:
      - application/json
      responses:
        "200":
          description: Successfully loaded contacts
          schema:
            items:
              $ref: '#/definitions/models.Contact'
            type: array
      summary: Get all contacts with message counts
  /contacts/timeline:
    get:
      consumes:
      - application/json
      description: Returns the daily number of messages per user.
      parameters:
      - description: ID of contact (CSV-Name)
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ContactTimeline'
            type: array
        "400":
          description: contact missing
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Sums up messages from each person in contact during one day
  /delete-zip:
    delete:
      consumes:
      - application/json
      description: Löscht ein angegebenes temporäres Verzeichnis rekursiv
      produces:
      - application/json
      responses:
        "200":
          description: Temp-Verzeichnis gelöscht
          schema:
            type: string
        "400":
          description: Temp-Verzeichnis nicht angegeben oder ungültig
          schema:
            type: string
        "500":
          description: Fehler beim Löschen
          schema:
            type: string
      summary: Löscht ein temporäres Verzeichnis
      tags:
      - upload
  /groups:
    get:
      description: Returns Groups and count of how many messages a user has
      produces:
      - application/json
      responses:
        "200":
          description: List of groups wiht message count
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
  /groups/timeline:
    get:
      consumes:
      - application/json
      description: Returns the daily number of messages per user for a group.
      parameters:
      - description: Name of group (CSV-Name)
        in: query
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GroupTimeline'
            type: array
        "400":
          description: group missing
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Sums up messages from each person in group during one day
  /upload-zip:
    post:
      consumes:
      - multipart/form-data
      description: Nimmt eine passwortgeschützte ZIP-Datei entgegen, entpackt sie
        in das relative Verzeichnis ./data
      parameters:
      - description: Passwort der ZIP-Datei
        in: header
        name: X-Zip-Password
        required: true
        type: string
      - description: ZIP-Datei
        in: formData
        name: file
        required: true
        type: file
      produces:
      - text/plain
      responses:
        "200":
          description: ZIP successfully extracted
          schema:
            type: string
        "400":
          description: Bad request / wrong password / invalid zip
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Upload ZIP (password protected)
      tags:
      - upload
  /users:
    get:
      description: Returns all identities from identity.csv
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Identity'
            type: array
  /wordcloud:
    get:
      description: |-
        Reads all CSV files named `message_*.csv` from `/data`, analyzes the `body` column,
        counts all words across all rows and returns the most frequent words as JSON.
      parameters:
      - description: Maximum number of words to return (default 100)
        in: query
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Most frequent words
          schema:
            items:
              $ref: '#/definitions/models.WordCount'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "404":
          description: No message files found
          schema:
            type: string
        "500":
          description: Internal server error while reading or parsing files
          schema:
            type: string
      summary: Get most frequent words
swagger: "2.0"